simple-rag-ai-agent-main/backend/requirements.txt
fastapi
uvicorn
pydantic
pypdf
tiktoken
faiss-cpu
numpy
openai
reportlab
python-multipart